requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.9
django-environ==0.8.1
django-heroku==0.3.1
django-markdownx==3.0.1
Markdown==3.3.4
Pillow==8.4.0
psycopg2==2.9.1
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
whitenoise==5.3.0
